chore: rename states to hooks
